requirements.txt
flask===1.1.2
requests
python3-wget
urllib3
